requirements.txt
numpy==1.24.3
Pillow==10.0.0
rawpy==0.17.1 